DemoActions

DemoActions.DemoActions

DemoActions.DP5Cache

DemoActions.bistable_gene_circuit!

DemoActions.dp5

DemoActions.dp5!

DemoActions.dp5_cache

DemoActions.dp5_inplace!

DemoActions.eulers_inplace_f!

DemoActions.lotka!

DemoActions.midpoint!

DemoActions.midpoint_inplace_f!

DemoActions.mysolve

DemoActions.mysolve_inplace

DemoActions.negative_autoregulation!

DemoActions.phosphorylation_cycle!

DemoActions.rk4!

DemoActions.rk4_inplace_f!

DemoActions.rk4_inplace_f_prealloc_k!

DemoActions.transcriptional_component!

DemoActions.DemoActions — Module

Learning repo for testing various CI and benchmarking tools.

Mainly benchmarking my naive ForwardDiff implementation and Dormand-Prince solver on models I find interesting.

Working through Mathematical Physiology by James Keener and Biomolecular Feedback Systems by Murray.

DemoActions.DP5Cache — Type

Dummy struct for the Dormand-Prince tableau

DemoActions.bistable_gene_circuit! — Method

Example 3.2 (Bistable gene circuit) - Murray

DemoActions.dp5! — Method

dp5!(m, f, u0, tspan, dt, p)

Explicit Dormand-Prince method. Fixed dt. Non-allocating for the solves. Allocates the intermediate f calls though.

DemoActions.dp5 — Method

dp5(f, u0, tspan, dt, p)

Uses dp5! to solve f and store in m.

DemoActions.dp5_cache — Method

dp5_cache()

allocates the DP5 tableau.

DemoActions.dp5_inplace! — Method

dp5_inplace!(m, f!, u0, tspan, dt, p)

Explicit Dormand-Prince method. Fixed dt.

DemoActions.eulers_inplace_f! — Method

eulers_inplace_f!(m, f!, u0, tspan, dt, p)

Non-allocating explicit first order method. Fixed dt.

DemoActions.lotka! — Method

lotka!(du, u, p, t)

lotka-volterra, inplace

DemoActions.midpoint! — Method

midpoint!(m, f, u0, tspan, dt, p)

Explicit second order method. Fixed dt. Allocates for intermediate k values. This means f does not return nothing.

DemoActions.midpoint_inplace_f! — Method

midpoint_inplace_f!(m, f, u0, tspan, dt, p)

Explicit second order method. Fixed dt.

TODO: have du allocated beforehand, so that this function has zero allocations.

DemoActions.mysolve — Method

mysolve(alg, f, u0, tspan, dt; p=nothing)

Allocating solve. The difference between mysolve and mysolve_inplace is mysolve expects f to return an array, not mutate one.

DemoActions.mysolve_inplace — Method

mysolve_inplace(alg!, f!, u0, tspan, dt; p=nothing)

Non-allocating solve. The difference between mysolve and mysolve_inplace is mysolve expects f to return an array, not mutate one.

DemoActions.negative_autoregulation! — Method

Example 3.4 (Negative autoregulation) - Murray

DemoActions.phosphorylation_cycle! — Method

Example 3.5 (Phosphorylation cycle) - Murray

Z + X -> Z + X* Y + X* -> Y + X X + X* = Xtot Ytot = c

Notice that in this example there is a fixed amount of Xtot and Ytot.

DemoActions.rk4! — Method

rk4!(m, f, u0, tspan, dt, p)

Explicit fourth-order Runge-Kutta method. Fixed dt. m is mutated inplace, but f is not an in-place function.

DemoActions.rk4_inplace_f! — Method

rk4_inplace_f!(m, f!, u0, tspan, dt, p)

Explicit fourth-order Runge-Kutta method. Fixed dt. m is mutated inplace and f is an in-place function.

DemoActions.rk4_inplace_f_prealloc_k! — Method

rk4_inplace_f_prealloc_k!(m, f!, u0, tspan, dt, p)

Explicit fourth-order Runge-Kutta method. Fixed dt. m is mutated inplace and f is an in-place function.

The difference with the other rk4 methods is that the k values are allocated before the loop.

I want to see which of these is fastest.

DemoActions.transcriptional_component! — Method

Example 3.1 (Transcriptional component)